Fix multiple splits in a single run
There was an issue where the way split index was the last item being split, which broken when we separated items into left/right, leading to overwritten values or "token not found" errors.

This stops us searching the run as soon as we find a place to split, and also guarantees that we don't try to start looking for a split past the end of the paragraph.

I've added a test that reproduces the issue, and then fixed the implementation such that the test passes.